+++ examples/resources/fmc_endpoint_device_types/import.sh
@@ -1,0 +1,1 @@
+terraform import fmc_endpoint_device_types.example "<domain>,[<endpoint_device_types_name>]"