+++ internal/provider/model_fmc_access_control_policy.go
@@ -85,6 +85,7 @@ type AccessControlPolicyRules struct {
 	DestinationPortLiterals    []AccessControlPolicyRulesDestinationPortLiterals    `tfsdk:"destination_port_literals"`
 	DestinationPortObjects     []AccessControlPolicyRulesDestinationPortObjects     `tfsdk:"destination_port_objects"`
 	SourceSgtObjects           []AccessControlPolicyRulesSourceSgtObjects           `tfsdk:"source_sgt_objects"`
+	EndpointDeviceTypes        []AccessControlPolicyRulesEndpointDeviceTypes        `tfsdk:"endpoint_device_types"`
 	DestinationSgtObjects      []AccessControlPolicyRulesDestinationSgtObjects      `tfsdk:"destination_sgt_objects"`
 	SourceZones                []AccessControlPolicyRulesSourceZones                `tfsdk:"source_zones"`
 	DestinationZones           []AccessControlPolicyRulesDestinationZones           `tfsdk:"destination_zones"`
@@ -157,6 +158,11 @@ type AccessControlPolicyRulesDestinationPortObjects struct {
 	Id types.String `tfsdk:"id"`
 }
 type AccessControlPolicyRulesSourceSgtObjects struct {
+	Name types.String `tfsdk:"name"`
+	Id   types.String `tfsdk:"id"`
+	Type types.String `tfsdk:"type"`
+}
+type AccessControlPolicyRulesEndpointDeviceTypes struct {
 	Name types.String `tfsdk:"name"`
 	Id   types.String `tfsdk:"id"`
 	Type types.String `tfsdk:"type"`
@@ -512,6 +518,22 @@ func (data AccessControlPolicy) toBody(ctx context.Context, state AccessControlP
 						itemChildBody, _ = sjson.Set(itemChildBody, "type", childItem.Type.ValueString())
 					}
 					itemBody, _ = sjson.SetRaw(itemBody, "sourceSecurityGroupTags.objects.-1", itemChildBody)
+				}
+			}
+			if len(item.EndpointDeviceTypes) > 0 {
+				itemBody, _ = sjson.Set(itemBody, "endPointDeviceTypes", []interface{}{})
+				for _, childItem := range item.EndpointDeviceTypes {
+					itemChildBody := ""
+					if !childItem.Name.IsNull() {
+						itemChildBody, _ = sjson.Set(itemChildBody, "name", childItem.Name.ValueString())
+					}
+					if !childItem.Id.IsNull() {
+						itemChildBody, _ = sjson.Set(itemChildBody, "id", childItem.Id.ValueString())
+					}
+					if !childItem.Type.IsNull() {
+						itemChildBody, _ = sjson.Set(itemChildBody, "type", childItem.Type.ValueString())
+					}
+					itemBody, _ = sjson.SetRaw(itemBody, "endPointDeviceTypes.-1", itemChildBody)
 				}
 			}
 			if len(item.DestinationSgtObjects) > 0 {
@@ -1082,6 +1104,30 @@ func (data *AccessControlPolicy) fromBody(ctx context.Context, res gjson.Result)
 						data.Type = types.StringNull()
 					}
 					(*parent).SourceSgtObjects = append((*parent).SourceSgtObjects, data)
+					return true
+				})
+			}
+			if value := res.Get("endPointDeviceTypes"); value.Exists() {
+				data.EndpointDeviceTypes = make([]AccessControlPolicyRulesEndpointDeviceTypes, 0)
+				value.ForEach(func(k, res gjson.Result) bool {
+					parent := &data
+					data := AccessControlPolicyRulesEndpointDeviceTypes{}
+					if value := res.Get("name"); value.Exists() {
+						data.Name = types.StringValue(value.String())
+					} else {
+						data.Name = types.StringNull()
+					}
+					if value := res.Get("id"); value.Exists() {
+						data.Id = types.StringValue(value.String())
+					} else {
+						data.Id = types.StringNull()
+					}
+					if value := res.Get("type"); value.Exists() {
+						data.Type = types.StringValue(value.String())
+					} else {
+						data.Type = types.StringNull()
+					}
+					(*parent).EndpointDeviceTypes = append((*parent).EndpointDeviceTypes, data)
 					return true
 				})
 			}
@@ -2131,6 +2177,59 @@ func (data *AccessControlPolicy) fromBodyPartial(ctx context.Context, res gjson.
 				data.Type = types.StringNull()
 			}
 			(*parent).SourceSgtObjects[i] = data
+		}
+		for i := 0; i < len(data.EndpointDeviceTypes); i++ {
+			keys := [...]string{"id"}
+			keyValues := [...]string{data.EndpointDeviceTypes[i].Id.ValueString()}
+
+			parent := &data
+			data := (*parent).EndpointDeviceTypes[i]
+			parentRes := &res
+			var res gjson.Result
+
+			parentRes.Get("endPointDeviceTypes").ForEach(
+				func(_, v gjson.Result) bool {
+					found := false
+					for ik := range keys {
+						if v.Get(keys[ik]).String() != keyValues[ik] {
+							found = false
+							break
+						}
+						found = true
+					}
+					if found {
+						res = v
+						return false
+					}
+					return true
+				},
+			)
+			if !res.Exists() {
+				tflog.Debug(ctx, fmt.Sprintf("removing EndpointDeviceTypes[%d] = %+v",
+					i,
+					(*parent).EndpointDeviceTypes[i],
+				))
+				(*parent).EndpointDeviceTypes = slices.Delete((*parent).EndpointDeviceTypes, i, i+1)
+				i--
+
+				continue
+			}
+			if value := res.Get("name"); value.Exists() && !data.Name.IsNull() {
+				data.Name = types.StringValue(value.String())
+			} else {
+				data.Name = types.StringNull()
+			}
+			if value := res.Get("id"); value.Exists() && !data.Id.IsNull() {
+				data.Id = types.StringValue(value.String())
+			} else {
+				data.Id = types.StringNull()
+			}
+			if value := res.Get("type"); value.Exists() && !data.Type.IsNull() {
+				data.Type = types.StringValue(value.String())
+			} else {
+				data.Type = types.StringNull()
+			}
+			(*parent).EndpointDeviceTypes[i] = data
 		}
 		for i := 0; i < len(data.DestinationSgtObjects); i++ {
 			keys := [...]string{"id"}